CRUD Enterprise ok

+++ ferreteriaThymeleaf/src/main/java/bigworkers/ingresoegreso/ferreteriaThymeleaf/service/DocumentTypeService.java
@@ -16,8 +16,8 @@ public class DocumentTypeService implements IDocumentTypeService{
     private IDocumentTypeRepository documentTypeRepository;
 
     @Override
-    public DocumentType findById(int id) {
-        Optional<DocumentType> tipoDocumento = documentTypeRepository.findById((long)id);
+    public DocumentType findById(long id) {
+        Optional<DocumentType> tipoDocumento = documentTypeRepository.findById(id);
         return tipoDocumento.get();
     }
 
@@ -34,13 +34,13 @@ public class DocumentTypeService implements IDocumentTypeService{
     }
 
     @Override
-    public DocumentType updateDocumentType(int id, DocumentType tipoDocumento) {
+    public DocumentType updateDocumentType(long id, DocumentType tipoDocumento) {
         DocumentType putTipoDocumento = documentTypeRepository.save(tipoDocumento);
         return putTipoDocumento;
     }
 
     @Override
-    public void deleteDocumentType(int id) {
-        documentTypeRepository.deleteById((long)id);
+    public void deleteDocumentType(long id) {
+        documentTypeRepository.deleteById(id);
     }
 }
