+++ ferreteriaThymeleaf/src/main/java/bigworkers/ingresoegreso/ferreteriaThymeleaf/service/ITransactionService.java
@@ -8,15 +8,15 @@ import java.util.List;
 
 public interface ITransactionService {
     //metodos abstractos
-    public Transaction findById(int id);
+    public Transaction findById(long id);
 
     public List<Transaction> findAll();
 
     public Transaction CreateTransaction (Transaction movimiento);
 
-    public Transaction updateTransaction(int id,Transaction movimiento);
+    public Transaction updateTransaction(long id,Transaction movimiento);
 
-    public void deleteTransaction(int id);
+    public void deleteTransaction(long id);
 
 
 }
