+++ ferreteriaThymeleaf/src/main/java/bigworkers/ingresoegreso/ferreteriaThymeleaf/service/ProfileService.java
@@ -16,8 +16,8 @@ public class ProfileService implements IProfileService{
     @Autowired
     private IProfileRepository profileRepository;
     @Override
-    public Profile findById(int id) {
-        Optional<Profile> perfil = profileRepository.findById((long)id);
+    public Profile findById(long id) {
+        Optional<Profile> perfil = profileRepository.findById(id);
         return perfil.get();
     }
 
@@ -34,13 +34,13 @@ public class ProfileService implements IProfileService{
     }
 
     @Override
-    public Profile updateProfile(int id, Profile perfil) {
+    public Profile updateProfile(long id, Profile perfil) {
         Profile putPerfil = profileRepository.save(perfil);
         return putPerfil;
     }
 
     @Override
-    public void deleteProfile(int id) {
-        profileRepository.deleteById((long)id);
+    public void deleteProfile(long id) {
+        profileRepository.deleteById(id);
     }
 }
